linter: Update rules to all static methods at end of class
The static methods are for assisting object persistence, not
directly related to the base functionality of the class.

In the future, I might move JSON helpers to a parent class, but
there are some nuances of object references that I don't have enough
understanding of yet. This works right now, and I don't want to break
it.