feature: Add `Project.handleEvent()` to handle click events
Passing `this` to `addEventListener()` automatically call
`this.handleEvent(event)`.

`event` contains info about the element that was clicked, and `this` is
used as context. This means I have the context of the instance of
`Project` that created the event listener, and properties of the element
that was clicked, which is all of the information needed to drive the UI
behavior.

refs #37